Align '.data' and '.rodata' sections to page size
If we don't do this, the linker will generate a warning, since the
writable '.data' section shouldn't be loaded with the '.text' section in
an executable page. This is not really an issue at this point, since we
aren't dealing with paging yet, but it's still annoying whenever the ELF
binaries are generated.

This commit shouldn't be an issue, but it is, because there is currently
a bug when loading a file that spans multiple clusters from the Stage 1
code. This bug is also related to a bug on my FAT dumping tool, which
checked if the absolute position in the FAT (in bytes) was odd or even,
when the 12-bit FAT index should have been used.